add tests for changing atomic idempotent

+++ src/meta/test/state_sync_test.cpp
@@ -153,12 +153,12 @@ void meta_service_test_app::state_sync_test()
             info.app_type = "simple_kv";
             info.app_name = "test_app" + boost::lexical_cast<std::string>(i);
             info.max_replica_count = 3;
-            info.partition_count = random32(100, 10000);
+            info.partition_count = static_cast<int32_t>(random32(100, 10000));
             info.status = dsn::app_status::AS_CREATING;
 
             // `atomic_idempotent` will be set true for the table with even index,
             // otherwise false.
-            info.atomic_idempotent = (i & 1) == 0;
+            info.atomic_idempotent = (static_cast<uint32_t>(i) & 1U) == 0;
 
             std::shared_ptr<app_state> app = app_state::create(info);
 
@@ -201,7 +201,7 @@ void meta_service_test_app::state_sync_test()
 
             // Recovered `app->atomic_idempotent` will true for the table with even index,
             // otherwise false.
-            ASSERT_EQ((i & 1) == 0, app->atomic_idempotent);
+            ASSERT_EQ((static_cast<uint32_t>(i) & 1) == 0, app->atomic_idempotent);
 
             for (int j = 0; j < app->partition_count; ++j) {
                 config_context &cc = app->helpers->contexts[j];
